# API-Engine-distro/API-Engine-distro-data/src/main/resources/local/swagger.yaml
swagger: '2.0'
info:
  version: v1
  title: API Engine
basePath: /API-Engine-web/v1
...
schemes:
  - http
paths:
...
  /plugins:
    get:
      tags:
        - plugins
      summary: List All Plugins
      description: This endpoint returns a list of all plugins that have been added to the
      operationId: list
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: A list of plugins.
          schema:
            type: array
            items:
              $ref: '#/definitions/PluginSummaryBean'
    post:
      tags:
        - plugins
      summary: Add a Plugin
      description: Use this endpoint to add a plugin to apiman.  A plugin consists of the maven coordinates of an artifact deployed to a remote maven repository (e.g. maven central).
      operationId: create
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/NewPluginBean'
      responses:
        '200':
          description: Full details about the plugin that was added.
          schema:
            $ref: '#/definitions/PluginBean'
  '/plugins/{pluginId}':
    get:
      tags:
        - plugins
      summary: Get Plugin by ID
      description: This endpoint can be used to access the full information about an apiman plugin.  The plugin is retrieved using the ID it was given when it was added.  The ID information can be retrieved by listing all plugins or remembered when a plugin is first added.
      operationId: get
      produces:
        - application/json
      parameters:
        - name: pluginId
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: An api-engine plugin.
          schema:
            $ref: '#/definitions/PluginBean'
    delete:
      tags:
        - plugins
      summary: Delete a Plugin by ID
      description: Call this endpoint to delete a plugin.
      operationId: delete
      parameters:
        - name: pluginId
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: 'successful, no content'
  '/plugins/{pluginId}/policyDefs':
    get:
      tags:
        - plugins
      summary: Get Plugin Policy Definitions
      description: Use this endpoint to get a list of all policy definitions contributed by the plugin.
      operationId: getPolicyDefs
      produces:
        - application/json
      parameters:
        - name: pluginId
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: A list of policy definitions.
          schema:
            type: array
            items:
              $ref: '#/definitions/PolicyDefinitionSummaryBean'
  '/plugins/{pluginId}/policyDefs/{policyDefId}/form':
    get:
      tags:
        - plugins
      summary: Get Plugin Policy Form
      description: 'Use this endpoint to retrieve the form associated with a particular policy definition.  Plugins may contribute policy definitions to apiman.  Part of that contribution *may* include a form for the UI to display when configuring an instance of the policy.  This endpoint returns this form.'
      operationId: getPolicyForm
      produces:
        - application/json
      parameters:
        - name: pluginId
          in: path
          required: true
          type: integer
          format: int64
        - name: policyDefId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: A policy configuration form.
          schema:
            type: string
...
definitions:
...
  PluginBean:
    type: object
    properties:
      id:
        type: integer
        format: int64
      groupId:
        type: string
      artifactId:
        type: string
      version:
        type: string
      classifier:
        type: string
      type:
        type: string
      name:
        type: string
      description:
        type: string
      createdBy:
        type: string
      createdOn:
        type: string
        format: date-time
...
  NewPluginBean:
    type: object
    properties:
      groupId:
        type: string
      artifactId:
        type: string
      version:
        type: string
      classifier:
        type: string
      type:
        type: string
      name:
        type: string
      description:
        type: string
...
  PolicyDefinitionSummaryBean:
    type: object
    properties:
      id:
        type: string
      policyImpl:
        type: string
      name:
        type: string
      description:
        type: string
      icon:
        type: string
      formType:
        type: string
        enum:
          - Default
          - JsonSchema
      pluginId:
        type: integer
        format: int64
...
  PluginSummaryBean:
    type: object
    properties:
      id:
        type: integer
        format: int64
      groupId:
        type: string
      artifactId:
        type: string
      version:
        type: string
      classifier:
        type: string
      type:
        type: string
      name:
        type: string
      description:
        type: string
      createdBy:
        type: string
      createdOn:
        type: string
        format: date-time
...